Change "distance spacing" multipler type to `double`
Avoids losing precision on initial load, causing an unnecessary
hash change in `EditorChangeHandler`.

Resolves test failures in `TestSceneEditorChangeStates`
(https://github.com/ppy/osu/runs/5192493482?check_suite_focus=true).

+++ osu.Game/Rulesets/Edit/IPositionSnapProvider.cs
@@ -14,7 +14,7 @@ namespace osu.Game.Rulesets.Edit
         /// The spacing multiplier applied to beat snap distances.
         /// </summary>
         /// <seealso cref="BeatmapInfo.DistanceSpacing"/>
-        IBindable<float> DistanceSpacingMultiplier { get; }
+        IBindable<double> DistanceSpacingMultiplier { get; }
 
         /// <summary>
         /// Given a position, find a valid time and position snap.
